Fixed some testsuite issues pointed out by Jim Galarowicz  - - - - - - - Fix issue where a 64-bit void * was cast directly to an int.  Now we cast it to a long, then to an int.

+++ testsuite/src/test13.mutatee.c
@@ -18,7 +18,7 @@ volatile int threads_running[NTHRD];
 
 void *init_func(void *arg)
 {
-   threads_running[(int) arg] = 1;
+   threads_running[(int) (long) arg] = 1;
    while (!done);
    return arg;
 }
